add PartitionedRegions
